chahua: 拆 persona_import.py 采集 / 写盘两层 → _persona_collect.py + _persona_fs.py
把纯文件采集（本地递归读 / GitHub 递归下载 + _MAX_* 限额 + _PackBudget）抽到
chahua/_persona_collect.py，把写盘 / 原子替换 / 崩溃恢复（_write_files /
_replace_dir_atomic / _recover_interrupted_updates）抽到 chahua/_persona_fs.py。
persona_import.py 留「import/update/check/delete」编排层 + 名字校验。

保住 monkeypatch 契约：被抽出的纯函数都不 bare-call 任何被 patch 的名字
（_gh_latest_commit_sha / _collect_local_files / _collect_from_source /
_safe_latest_commit_sha / _gh_fetch_one_file），故搬走后语义不变；其 bare-name
调用方（_check_github / _check_folder / _detect_local_modified / _collect_from_source
/ import_from_* / update_persona）全留在 persona_import，经本模块命名空间解析。
_collect_local_files / _fetch_github_dir / _write_files / _replace_dir_atomic /
_recover_interrupted_updates re-import 回 persona_import，公开 import 路径与
patch 目标不变。承重不变量（原子 swap / 防穿越 / 限额 / 崩溃恢复）docstring 原样迁移。

persona_import.py 885 → 633 行。
